luci-mod-network: include prefix in interface name check
The interface name check only considered the user input when validating
the length, which could result in the actual name exceeding the kernel
limit of 15 characters when the bridge/protocol prefix was included as
well. For example the interface name abcdefgh1234567 was accepted as
valid but with prefixes included (such as br-abcdefgh1234567 or
gre-abcdefgh1234567) it exceeded the limit.

Signed-off-by: Daniel Nilsson <dannil+github@protonmail.com>